refactor(web): split types.ts into types, routing, and projects
types.ts was a 530-line grab bag mixing data interfaces, URL routing,
project matching, and hub topology. Split into focused modules:

- types.ts (75 lines): pure data interfaces, zero functions
- routing.ts (184 lines): URL parsing, view resolution, path building
- projects.ts (290 lines): project matching, sidebar folders, topology,
  normalizeRemote (moved from types.ts, its only consumer)

No behavior changes.
